% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_utils.R
\name{assign2center}
\alias{assign2center}
\title{Assign a point to the cell with the closest center}
\usage{
assign2center(distance.vector)
}
\arguments{
\item{distance.vector}{vector of length K containing the distances between the point and the centers}
}
\value{
it returns the index(es) of the cell(s) with the closest center(s). Returns 0 it the minimum does not exists
}
\description{
...
}
\details{
...
}
