% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_bootstrapVar.R
\name{boot.var.man}
\alias{boot.var.man}
\title{Bootstrap variance for method mixed_RDD}
\usage{
boot.var.man(res_RDD_OOK, K, metric_manifold)
}
\description{
Bootstrap variance for method mixed_RDD
}
