% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_utils.R
\name{create.rdd}
\alias{create.rdd}
\title{Divide the domain in K subregions}
\usage{
create.rdd(K, method.rdd = "Voronoi", data_coords, graph.distance,
  nk_min, grid, data.grid.distance, suppressMes = T)
}
\arguments{
\item{K}{number of regions the domain is divided in}

\item{method.rdd}{method used to define the subregions. So far only "Voronoi" has been implemented}

\item{data_coords}{\code{N*2} or \code{N*3} matrix of [lat,long], [x,y] or [x,y,z] coordinates of the locations where data has been measured}

\item{graph.distance}{\code{N*N} distance matrix (the [i,j] element is the length of the shortest path between points i and j)}

\item{nk_min}{minimum number of observations within a cell}

\item{grid}{prediction grid, i.e. \code{M*2} or \code{M*3} matrix of coordinates where to predict}

\item{data.grid.distance}{\code{N*M} distance matrix between locations where the datum has been observed and locations where
the datum has to be predicted}

\item{suppressMes}{\{\code{TRUE}, \code{FALSE}\} controls the level of interaction and warnings given}
}
\value{
it returns a list with the following fields
\itemize{
      \item \code{assign} {vector of length N that indicates, for every known location, the cell it has been assigned to}
      \item \code{centers} {\code{K*3} matrix reporting the coordinates and the index of the locations drawn as centers of the K cells }
      \item \code{assigng} {vector of length M that indicates, for every new location, the cell it has been assigned to}
      \item \code{gridk} {list of K elements. The i-th element contains the coordinates of the grid points assigned to the i-th cell}
      \item \code{graph.distance.grid.centers} {\code{K*M} matrix containing the distances between each grid points and the K centers}
}
}
\description{
...
}
\details{
...
}
