% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_utils.R
\name{Eucldist}
\alias{Eucldist}
\title{Compute the Euclidean distance between two points}
\usage{
Eucldist(c1, c2)
}
\arguments{
\item{c1}{coordinates of the first point.}

\item{c2}{coordinates of the second point.}
}
\value{
the Euclidean distance between \code{c1} and \code{c2}
}
\description{
...
}
\details{
...
}
