% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_RDD.R
\name{full_RDD}
\alias{full_RDD}
\title{Perform full_RDD}
\usage{
full_RDD(data_coords, data_val, K, grid, nk_min = 1, B = 100,
  suppressMes = F, tol = 1e-12, max_it = 100, n_h = 15,
  tolerance_intrinsic = 10^(-6), X = NULL, X_new = NULL,
  ker.width.intrinsic = 0, ker.width.vario = 1.5,
  graph.distance.complete, data.grid.distance, aggregation_mean,
  aggregation_kriging, method.analysis = "Local mean", metric_manifold,
  metric_ts, model_ts, vario_model, distance = NULL)
}
\arguments{
\item{data_coords}{\code{N*2} or \code{N*3} matrix of [lat,long], [x,y] or [x,y,z] coordinates. [lat,long] are supposed to
be provided in signed decimal degrees}

\item{data_val}{array [\code{p,p,N}] of \code{N} symmetric positive definite matrices of dimension \code{p*p}}

\item{K}{number of cells the domain is subdivided in}

\item{grid}{prediction grid, i.e. \code{M*2} or \code{M*3} matrix of coordinates where to predict}

\item{nk_min}{minimum number of observations within a cell}

\item{B}{number of \emph{divide} iterations to perform}

\item{suppressMes}{\{\code{TRUE}, \code{FALSE}\} controls the level of interaction and warnings given}

\item{tol}{tolerance for the main loop of \code{model_kriging}}

\item{max_it}{maximum number of iterations for the main loop of \code{model_kriging}}

\item{n_h}{number of bins in the empirical variogram}

\item{tolerance_intrinsic}{tolerance for the computation of the intrinsic mean}

\item{X}{matrix (N rows and unrestricted number of columns) of additional covariates for the tangent space model, possibly NULL}

\item{X_new}{matrix (with the same number of rows of \code{new_coords}) of additional covariates for the new locations, possibly NULL}

\item{ker.width.intrinsic}{parameter controlling the width of the Gaussian kernel for the computation of the local mean (if 0, 
a "step kernel" is used, giving weight 1 to all the data within the cell and 0 to those outside of it)}

\item{ker.width.vario}{parameter controlling the width of the Gaussian kernel for the computation of the empirical variogram (if 0, 
a "step kernel" is used, giving weight 1 to all the data within the cell and 0 to those outside of it)}

\item{graph.distance.complete}{\code{N*N} distance matrix (the [i,j] element is the length of the shortest path between points i and j)}

\item{data.grid.distance}{\code{N*M} distance matrix between locations where the datum has been observed and locations where
the datum has to be predicted}

\item{aggregation_mean}{"Weighted" to aggregate the mean predictions using kernel-based weights, "Equal" to use equal weights}

\item{aggregation_kriging}{"Weighted" to aggregate the Kriging predictions using kernel-based weights, "Equal" to use equal weights}

\item{method.analysis}{"Local mean" to predict just with the mean, "Kriging" to predict via Kriging procedure}

\item{metric_manifold}{metric used on the manifold. It must be chosen among "Frobenius", "LogEuclidean", "SquareRoot"}

\item{metric_ts}{metric used on the tangent space. It must be chosen among "Frobenius", "FrobeniusScaled", "Correlation"}

\item{model_ts}{type of model fitted on the tangent space. It must be chosen among "Intercept", "Coord1", "Coord2", "Additive"}

\item{vario_model}{type of variogram fitted. It must be chosen among "Gaussian", "Spherical", "Exponential"}

\item{distance}{type of distance between coordinates. It must be either "Eucldist" or "Geodist"}
}
\value{
According to the analysis chosen:
\itemize{
  \item If \code{method.analysis} = "Local mean" it returns a list with the following fields
    \itemize{
      \item \code{resBootstrap} {A list consisting of}
         \itemize{
          \item \code{fmean} {list of length \code{B}. Each field contains the prediction (at iteration \code{b}) for each new location, obtained 
                            as the intrinsic mean of the data within the tile it belongs to}
          \item \code{kervalues_mean} {Weights used for aggregating \code{fmean}}
         }
      \item \code{resAggregated} {Predictions, for each new location, obtained aggregating \code{fmean} using \code{kervalues_mean} as weights}
    }
  \item If \code{method.analysis} = "Kriging" it returns a list with the following fields
    \itemize{
      \item \code{resBootstrap} {A list consisting of}
         \itemize{
          \item \code{fmean} {list of length \code{B}. Each field contains the prediction (at iteration \code{b}) for each new location, obtained 
                            as the intrinsic mean of the data within the tile it belongs to}
          \item \code{fpred} {list of length \code{B}. Each field contains the prediction (at iteration \code{b}) for each new location, obtained 
                            through kriging}
          \item \code{kervalues_mean} {Weights used for aggregating \code{fmean}}
          \item \code{kervalues_krig} {Weights used for aggregating \code{fpred}}
          \item \code{variofit} {list of length \code{B}. Each field contains, for each datum, the parameters of the variogram fitted in the tile it belongs to}
         }
      \item \code{resAggregated} {Predictions, for each new location, obtained aggregating \code{fpred} using \code{kervalues_krig} as weights}
      \item \code{resLocalMean} {Predictions, for each new location, obtained aggregating \code{fmean} using \code{kervalues_mean} as weights}
    }
}
}
\description{
Perform kriging prediction using FullRDD procedure
}
\details{
It uses a repetition of local analyses, through a \emph{divide} et \emph{impera} strategy. In the \emph{divide} step, the domain
is randomly decomposed in subdomains where local tangent-space models are estimated in order to predict at new locations (in each subregion is 
performed exactly the analysis described in the \code{model_kriging} function). This is repeated \code{B} times with different 
partitions of the domain. Then, in the \emph{impera} step, the results of these iterations are aggregated, by means of the intrinsic mean,
to provide a final prediction.
}
