% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_RDD_bootstrapVar.R
\name{full_RDD_bootstrapVar}
\alias{full_RDD_bootstrapVar}
\title{Bootstrap variance for method full_RDD}
\usage{
full_RDD_bootstrapVar(res_RDD_OOK, K, n, metric_manifold, dim_matrix)
}
\description{
Bootstrap variance for method full_RDD
}
