% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_utils.R
\name{Geodist}
\alias{Geodist}
\title{Compute the great-circle distance between two points}
\usage{
Geodist(c1, c2)
}
\arguments{
\item{c1}{coordinates [lat, long] of the first point.}

\item{c2}{coordinates [lat, long] of the second point.}
}
\value{
the great-circle distance between \code{c1} and \code{c2}
}
\description{
...
}
\details{
The distance is computed using the Haversine formula
}
