% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinsic_mean.R
\name{intrinsic_mean}
\alias{intrinsic_mean}
\title{Intrinsic mean}
\usage{
intrinsic_mean(data, metric_manifold = "Frobenius",
  metric_ts = "Frobenius", tolerance = 1e-06, weight = NULL)
}
\arguments{
\item{data}{list or array [\code{n,n,B}] of \code{B} symmetric positive definite matrices of dimension \code{nxn}}

\item{metric_manifold}{metric used on the manifold. It must be chosen among "Frobenius", "LogEuclidean", "SquareRoot"}

\item{metric_ts}{metric used on the tangent space. It must be either "Frobenius" or "FrobeniusScaled"}

\item{tolerance}{tolerance for the computation of the intrinsic_mean}

\item{weight}{vector of length \code{N} to weight the locations in the computation of the intrinsic mean. If NULL
a vector of ones is used}
}
\description{
Evaluate the intrinsic mean of a given set of symmetric positive definite matrices
}
\details{
...
}
\examples{
data_manifold_tot <- Manifoldgstat::fieldCov
Sigma <-intrinsic_mean(data_manifold_tot, metric_manifold = "Frobenius", metric_ts = "Frobenius")
print(Sigma)
}
