% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_utils.R
\name{kerfn}
\alias{kerfn}
\title{Evaluate a gaussian kernel}
\usage{
kerfn(newdata, center, ker.type = "Gau", param)
}
\arguments{
\item{newdata}{coordinates of the locations where we want to compute the kernel values}

\item{center}{coordinates of the reference center for the kernel}

\item{ker.type}{type of kernel. So far only "Gau" (i.e gaussian kernel) has been implemented}

\item{param}{parameters that define the kernel. For the gaussian kernel it is the sigma parameter}
}
\value{
the values of the kernel function corresponding to the vector of Euclidean distances between \code{newdata} and the \code{center}
}
\description{
...
}
\details{
...
}
