% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{kriging}
\alias{kriging}
\title{Kriging}
\usage{
kriging(GLS_model, Sigma, coords, new_coords, model_ts = "additive",
  vario_model = "Gaussian", metric_manifold = "Frobenius",
  X_new = NULL, distance = "Geodist")
}
\description{
Kriging
}
