% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_bootstrapVar.R
\name{mixed_bootstrapVar}
\alias{mixed_bootstrapVar}
\title{Bootstrap variance for method mixed_RDD}
\usage{
mixed_bootstrapVar(res_RDD_OOK, K, metric_manifold)
}
\description{
Bootstrap variance for method mixed_RDD
}
