% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_GLS_sigma_fixed.R
\name{model_GLS_sigma_fixed}
\alias{model_GLS_sigma_fixed}
\title{Model given the data}
\usage{
model_GLS_sigma_fixed(data_manifold, coords, X = NULL, Sigma,
  metric_manifold = "Frobenius", metric_ts = "Frobenius",
  model_ts = "additive", vario_model = "Gaussian", n_h = 15,
  distance = "Geodist", max_it = 100, tolerance = 10^(-4),
  weight = NULL)
}
\description{
Model given the data
}
