% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_krigingMIXED.R
\name{model_kriging_mixed}
\alias{model_kriging_mixed}
\title{Perform main routine for mixed_RDD}
\usage{
model_kriging_mixed(data_manifold, coords, X = NULL, Sigma_data,
  metric_manifold = "Frobenius", model_ts = "Additive",
  vario_model = "Gaussian", n_h = 15, distance = NULL,
  data_dist_mat = NULL, data_grid_dist_mat = NULL, max_it = 100,
  tolerance = 1e-06, max_sill = NULL, max_a = NULL, new_coords,
  Sigma_new, X_new = NULL, create_pdf_vario = TRUE,
  pdf_parameters = NULL, suppressMes = FALSE)
}
\arguments{
\item{data_manifold}{array [\code{p,p,N}] of \code{N} symmetric positive definite matrices of dimension \code{p*p}}

\item{coords}{\code{N*2} or \code{N*3} matrix of [lat,long], [x,y] or [x,y,z] coordinates.}

\item{X}{matrix (N rows and unrestricted number of columns) of additional covariates for the tangent space model, possibly NULL}

\item{Sigma_data}{List of the \code{N} fictional tangent points in correspondence with the data used to build the model}

\item{metric_manifold}{metric used on the manifold. It must be chosen among "Frobenius", "LogEuclidean", "SquareRoot"}

\item{model_ts}{type of model fitted on the tangent space. It must be chosen among "Intercept", "Coord1", "Coord2", "Additive"}

\item{vario_model}{type of variogram fitted. It must be chosen among "Gaussian", "Spherical", "Exponential"}

\item{n_h}{number of bins in the empirical variogram}

\item{distance}{type of distance between coordinates. It must be either "Eucldist" or "Geodist"}

\item{data_dist_mat}{\code{N*N} distance matrix (the [i,j] element is the length of the shortest path between points i and j)}

\item{data_grid_dist_mat}{\code{N*M} distance matrix between locations where the datum has been observed and locations where
the datum has to be predicted}

\item{max_it}{maximum number of iterations for the main loop of \code{model_kriging}}

\item{tolerance}{tolerance for the main loop of \code{model_kriging}}

\item{max_sill}{max value allowed for \code{sill} in the fitted variogram. If NULL it is defined as \code{1.15*max(emp_vario_values)}}

\item{max_a}{maximum value for \code{a} in the fitted variogram. If NULL it is defined as \code{1.15*h_max}}

\item{new_coords}{prediction grid, i.e. \code{M*2} or \code{M*3} matrix of coordinates where to predict}

\item{Sigma_new}{List of the \code{M} fictional tangent points in correspondence with the locations where we want to predict}

\item{X_new}{matrix (with the same number of rows of \code{new_coords}) of additional covariates for the new locations, possibly NULL}

\item{create_pdf_vario}{boolean. If \code{TRUE} the empirical and fitted variograms are plotted in a pdf file}

\item{pdf_parameters}{list with the fields \code{test_nr} and \code{sample_draw}. Additional parameters to name the pdf}

\item{suppressMes}{\{\code{TRUE}, \code{FALSE}\} controls the level of interaction and warnings given}
}
\value{
it returns a list with the following fields
\item{\code{beta}}{ vector of the beta matrices of the fitted model}
\item{\code{gamma_matrix}}{ \code{N*N} covariogram matrix}
\item{\code{residuals}}{ vector of the \code{N} residual matrices}
\item{\code{emp_vario_values}}{ vector of empircal variogram values in correspondence of \code{h_vec}}
\item{\code{h_vec}}{ vector of positions at which the empirical variogram is computed}
\item{\code{fitted_par_vario}}{ estimates of \emph{nugget}, \emph{sill-nugget} and \emph{practical range}}
\item{\code{iterations}}{ number of iterations of the main loop}
\item{\code{prediction}}{ vector of matrices predicted at the new locations}
}
\description{
...
}
\details{
...
}
