% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ker.R
\name{plot_ker_rect}
\alias{plot_ker_rect}
\title{Plot kernel}
\usage{
plot_ker_rect(data_coords, id, xmax, ymax, m, n, ker.width)
}
\arguments{
\item{data_coords}{coordinates of the data}

\item{id}{the index of the row of \code{data_coords} that will be used as center}

\item{xmax}{the maximum value for the x-coordinate (the minimum is 0)}

\item{ymax}{the maximum value for the y-coordinate (the minimum is 0)}

\item{m}{number of points on the grid in horizontal direction}

\item{n}{number of points on the grid in vertical direction}

\item{ker.width}{kernel width}
}
\description{
Plot kernel
}
