% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variogram.R
\name{plot_variogram}
\alias{plot_variogram}
\title{Plot empirical and fitted variogram}
\usage{
plot_variogram(empirical_variogram, fitted_variogram, model, distance)
}
\arguments{
\item{empirical_variogram}{A list containing the two following fields: 
- h_vec: vector of positions at which the empirical variogram is computed
- emp_vario_values: vector of empircal variogram values in correspondence of \code{h_vec}}

\item{fitted_variogram}{A list containing the two following fields: 
- hh: dense vector of positions at which \code{fit_vario_values} is computed
- fit_vario_values: Vector of fitted variogram values in correspondence of \code{hh}}

\item{model}{Type of variogram used for fitting (it will be reported on the y-axis). It can be "Gaussian", "Spherical" or "Exponential"}

\item{distance}{Type of distance used to compute \code{h_vec} (it will be reported on the x-axis). It must be either "Eucldist" or "Geodist"}
}
\description{
Plot empirical and fitted variogram
}
