% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_utils.R
\name{RDD_OOK_aggr_man}
\alias{RDD_OOK_aggr_man}
\title{Aggregate the results of the bootrap iterations}
\usage{
RDD_OOK_aggr_man(fOKBV, weights_intrinsic, ker.width.intrinsic)
}
\arguments{
\item{fOKBV}{list of length \code{B}, containing the results obtained, for each location, at the \code{B} bootstrap iterations
(Usually it is the \code{fmean} or \code{fpred} returned by \code{RDD_OOK_boot_man} or \code{RDD_OOK_boot_man_mixed})}

\item{weights_intrinsic}{weights to use to aggregate the results}

\item{ker.width.intrinsic}{width of the kernel used to compute \code{weights_intrinsic}. 0 if we use equal weights}
}
\value{
the aggregated result, obtained as the intrinsic_mean of \code{fOKBV}
}
\description{
...
}
\details{
If \code{ker.width.intrinsic!=0} the data are aggregated using the normalized \code{weights_intrinsic}. Otherwise equal weights are used
}
