% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDD_utils.R
\name{RDD_OOK_boot_man}
\alias{RDD_OOK_boot_man}
\title{Main routine for full_RDD}
\usage{
RDD_OOK_boot_man(data_coords, data_val, K, grid, nk_min, B, suppressMes,
  tol, max_it, n_h, tolerance_intrinsic, X, X_new, ker.width.intrinsic,
  ker.width.vario, graph.distance.complete, data.grid.distance,
  method.analysis, metric_manifold, metric_ts, model_ts, vario_model,
  distance)
}
\arguments{
\item{data_coords}{\code{N*2} or \code{N*3} matrix of [lat,long], [x,y] or [x,y,z] coordinates.}

\item{data_val}{list of \code{N} symmetric positive definite matrices of dimension \code{p*p}}

\item{K}{Number of neighborhood (i.e., centers) to sample at each iteration}

\item{grid}{\code{M*2} or \code{M*3} matrix of [lat,long], [x,y] or [x,y,z] coordinates of the new locations where to predict}

\item{nk_min}{Minimum number of observations within a neighborhood}

\item{B}{Number of bootstap iterations}

\item{suppressMes}{boolean. If \code{TRUE} warning messagges are not printed}

\item{tol}{tolerance for each main loop}

\item{max_it}{max number of iterations for each main loop}

\item{n_h}{number of bins in the emprical variogram}

\item{tolerance_intrinsic}{tolerance for the computation of the intrinsic mean}

\item{X}{Additional covariates for the locations used to create the model}

\item{X_new}{Additional covariates for the \code{M} locations where to perform kriging}

\item{ker.width.intrinsic}{Parameter controlling the width of the Gaussian kernel for the computation of the local mean (if 0, no kernel is used)}

\item{ker.width.vario}{Parameter controlling the width of the Gaussian kernel for the computation of the empirical variogram (if 0, no kernel is used)}

\item{graph.distance.complete}{\code{N*N} distance matrix (the [i,j] element is the length of the shortest path between points i and j)}

\item{data.grid.distance}{\code{N*M} distance matrix between locations where the datum has been observed and locations where
the datum has to be predicted}

\item{method.analysis}{"Local mean" to predict just with the mean, "Kriging" to predict via Kriging procedure}

\item{metric_manifold}{Metric used on the manifold. It must be chosen among "Frobenius", "LogEuclidean", "SquareRoot" and "Correlation"}

\item{metric_ts}{Metric used on the tangent space. It must be chosen among "Frobenius", "FrobeniusScaled", "Correlation"}

\item{model_ts}{Type of model fitted on the tangent space. It must be chosen among "Intercept", "Coord1", "Coord2", "Additive"}

\item{vario_model}{Type of variogram fitted. It must be chosen among "Gaussian", "Spherical", "Exponential"}

\item{distance}{Type of distance between coordinates. It must be either "Eucldist" or "Geodist"}
}
\value{
According to the analysis chosen:
\itemize{
  \item If \code{method.analysis} = "Local mean" it returns a list with the following fields
    \itemize{
          \item \code{fmean} {list of length \code{B}. Each field contains the prediction (at iteration \code{b}) for each new location, obtained
                            as the intrinsic mean of the data within the tile it belongs to}
          \item \code{kervalues_mean} {Weights used for aggregating \code{fmean}}
         }
  \item If \code{method.analysis} = "Kriging" it returns a list with the following fields
    \itemize{
          \item \code{fmean} {list of length \code{B}. Each field contains the prediction (at iteration \code{b}) for each new location, obtained
                            as the intrinsic mean of the data within the tile it belongs to}
          \item \code{fpred} {list of length \code{B}. Each field contains the prediction (at iteration \code{b}) for each new location, obtained
                            through kriging}
          \item \code{kervalues_mean} {Weights used for aggregating \code{fmean}}
          \item \code{kervalues_krig} {Weights used for aggregating \code{fpred}}
          \item \code{variofit} {list of length \code{B}. Each field contains, for each datum, the parameters of the variogram fitted in the tile it belongs to}
         }
  }
}
\description{
...
}
\details{
...
}
