% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_utils.R
\name{RDD_OOK_boot_man_mixed}
\alias{RDD_OOK_boot_man_mixed}
\title{Main routine for mixed_RDD}
\usage{
RDD_OOK_boot_man_mixed(data_coords, data_val, K, grid, nk_min, B,
  suppressMes, ker.width.intrinsic, graph.distance.complete,
  data.grid.distance, metric_ts, vario_model, tol, max_it, n_h,
  tolerance_intrinsic, X, X_new, metric_manifold, model_ts, distance)
}
\arguments{
\item{data_coords}{\code{N*2} or \code{N*3} matrix of [lat,long], [x,y] or [x,y,z] coordinates. [lat,long] are supposed to
be provided in signed decimal degrees}

\item{data_val}{array [\code{p,p,N}] of \code{N} symmetric positive definite matrices of dimension \code{p*p}}

\item{K}{number of cells the domain is subdivided in}

\item{grid}{prediction grid, i.e. \code{M*2} or \code{M*3} matrix of coordinates where to predict}

\item{nk_min}{minimum number of observations within a cell}

\item{B}{number of \emph{divide} iterations to perform}

\item{suppressMes}{\{\code{TRUE}, \code{FALSE}\} controls the level of interaction and warnings given}

\item{ker.width.intrinsic}{parameter controlling the width of the Gaussian kernel for the computation of the local mean (if 0, 
a "step kernel" is used, giving weight 1 to all the data within the cell and 0 to those outside of it)}

\item{graph.distance.complete}{\code{N*N} distance matrix (the [i,j] element is the length of the shortest path between points i and j)}

\item{data.grid.distance}{\code{N*M} distance matrix between locations where the datum has been observed and locations where
the datum has to be predicted}

\item{metric_ts}{metric used on the tangent space. It must be chosen among "Frobenius", "FrobeniusScaled", "Correlation"}

\item{vario_model}{type of variogram fitted. It must be chosen among "Gaussian", "Spherical", "Exponential"}

\item{tol}{tolerance for the main loop of \code{model_kriging}}

\item{max_it}{maximum number of iterations for the main loop of \code{model_kriging}}

\item{n_h}{number of bins in the empirical variogram}

\item{tolerance_intrinsic}{tolerance for the computation of the intrinsic mean}

\item{X}{matrix (N rows and unrestricted number of columns) of additional covariates for the tangent space model, possibly NULL}

\item{X_new}{matrix (with the same number of rows of \code{new_coords}) of additional covariates for the new locations, possibly NULL}

\item{metric_manifold}{metric used on the manifold. It must be chosen among "Frobenius", "LogEuclidean", "SquareRoot"}

\item{model_ts}{type of model fitted on the tangent space. It must be chosen among "Intercept", "Coord1", "Coord2", "Additive"}

\item{distance}{type of distance between coordinates. It must be either "Eucldist" or "Geodist"}
}
\value{
it returns a list with the following fields
\itemize{
      \item \code{fmean} {list of length \code{B}. Each field contains the prediction (at iteration \code{b}) for each location, obtained
                            as the intrinsic mean of the data within the tile it belongs to}
      \item \code{kervalues_mean} {Weights used for aggregating \code{fmean}}
      }
}
\description{
...
}
\details{
...
}
