% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_utils.R
\name{return_ith_list_element}
\alias{return_ith_list_element}
\title{Return a given element of a list}
\usage{
return_ith_list_element(lista, i)
}
\arguments{
\item{lista}{A list}

\item{i}{The index of the element to extract}
}
\value{
It returns the i-th element of \code{lista}
}
\description{
Return a given element of a list
}
