% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDD_utils.R
\name{return_ith_row}
\alias{return_ith_row}
\title{Return a given row of a matrix}
\usage{
return_ith_row(mat, i)
}
\arguments{
\item{mat}{A matrix}

\item{i}{The index of the row to extract}
}
\value{
It returns the i-th row of \code{mat}
}
\description{
Return a given row of a matrix
}
