% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapVar.R
\name{bootstrapVar}
\alias{bootstrapVar}
\title{Compute the bootstrap variance}
\usage{
bootstrapVar(res.boot, res.aggr, K, metric_manifold)
}
\arguments{
\item{res.boot}{A list of length \code{B}. Each field contains a list with the M predictions generated by the corresponding iteration}

\item{res.aggr}{A list of lenght M. Each field a single prediction, computed aggregating the corresponding data in res.boot}

\item{K}{number of cells the domain is subdivided in}

\item{metric_manifold}{metric used on the manifold. It must be chosen among "Frobenius", "LogEuclidean", "SquareRoot"}
}
\value{
It returns a vector of length M (i.e. the number of locations where we predict), containing the the prediction variance in the corresponding location
}
\description{
Compute the bootstrap variance
}
