% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_manifold.R
\name{distance_manifold}
\alias{distance_manifold}
\title{Distance on the manifold}
\usage{
distance_manifold(data1, data2, metric_manifold = "Frobenius")
}
\arguments{
\item{data1}{list or array [\code{p,p,B1}] of \code{B1} symmetric positive definite matrices of dimension \code{p*p}. Or a single \code{p*p} matrix.}

\item{data2}{list or array [\code{p,p,B2}] of \code{B2} symmetric positive definite matrices of dimension \code{p*p}. Or a single \code{p*p} matrix.}

\item{metric_manifold}{metric used on the manifold. It must be chosen among "Frobenius", "LogEuclidean", "SquareRoot"}
}
\value{
A double or a vector of distances
}
\description{
Compute the manifold distance between symmetric positive definite matrices
}
\details{
If \code{B2}=\code{B1} then the result is a vector of length \code{B1=B2} containing in position \code{i} the manifold distance between \code{data1[,,i]} and \code{data2[,,i]}.
Instead if \code{B2}=1 and \code{B1}!=1 the result is a vector of length \code{B1} containing in position \code{i} the manifold distance between \code{data1[,,i]} and \code{data2[,,1]}
}
\examples{
data_manifold_model <- Manifoldgstat::rCov
distances <-distance_manifold(data_manifold_model, diag(2), metric_manifold = "Frobenius")
print(distances)
}
