% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_manifold.R
\name{distance_manifold}
\alias{distance_manifold}
\title{Distance on the manifold}
\usage{
distance_manifold(data1, data2, metric_manifold = "Frobenius",
  metric_ts = "Frobenius")
}
\arguments{
\item{data1}{list or array [\code{n,n,B}] of \code{B} symmetric positive definite matrices of dimension \code{nxn}}

\item{data2}{can be a list or array [\code{n,n,B}] of \code{B} symmetric positive definite matrices of dimension \code{nxn}. 
Or it can be a single \code{nxn} matrix.}

\item{metric_manifold}{metric used on the manifold. It must be chosen among "Frobenius", "LogEuclidean", "SquareRoot"}
}
\description{
Compute the manifold distance between symmetric positive definite matrices
}
\details{
If \code{B2}=\code{B1} then the result is a vector of length \code{B1=B2} containing in position \code{i} the manifold distance beetween \code{data1[,,i]} and \code{data2[,,i]}.
Instead if \code{B2}=1 the result is a vector of length \code{B1} containing in position \code{i} the manifold distance between \code{data1[,,i]} and \code{data2[,,1]}
}
\examples{
data_manifold_model <- Manifoldgstat::rCov
distances <-distance_manifold(data_manifold_model, diag(2), metric_manifold = "Frobenius")
print(distances)
}
