% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinsic_mean.R
\name{intrinsic_mean}
\alias{intrinsic_mean}
\title{Intrinsic mean}
\usage{
intrinsic_mean(data, metric_manifold = "Frobenius",
  metric_ts = "Frobenius", tolerance = 1e-06,
  weight_intrinsic = NULL, weight_extrinsic = weight_intrinsic,
  tolerance_map_cor = 1e-06)
}
\arguments{
\item{data}{list or array [\code{n,n,B}] of \code{B} symmetric positive definite matrices of dimension \code{nxn}}

\item{metric_manifold}{metric used on the manifold. It must be chosen among "Frobenius", "LogEuclidean", "SquareRoot", "Correlation"}

\item{metric_ts}{metric used on the tangent space. It must be chosen among "Frobenius", "FrobeniusScaled", "Correlation"}

\item{tolerance}{tolerance for the computation of the intrinsic_mean}

\item{weight_intrinsic}{vector of length \code{B} to weight the matrices in the computation of the intrinsic mean. If NULL
a vector of ones is used}

\item{weight_extrinsic}{vector of length \code{B} to weight the matrices in the computation of the extrinsic mean. If NULL
weight_intrinsic are used}

\item{tolerance_map_cor}{tolerance to use in the maps. Required only if \code{metric_manifold== "Correlation"}}
}
\value{
A matrix representing the intrinsic mean of the \code{data}
}
\description{
Evaluate the intrinsic mean of a given set of symmetric positive definite matrices
}
\examples{
data_manifold_tot <- Manifoldgstat::fieldCov
Sigma <-intrinsic_mean(data_manifold_tot, metric_manifold = "Frobenius", metric_ts = "Frobenius")
print(Sigma)
}
