% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_RDD.R
\name{mixed_RDD}
\alias{mixed_RDD}
\title{Perform mixed_RDD}
\usage{
mixed_RDD(data_coords, data_val, K, grid, nk_min = 1, B = 100,
  suppressMes = F, ker.width.intrinsic = 0, graph.distance.complete,
  data.grid.distance, N_samples, p, aggregation_mean, metric_ts,
  tol = 1e-12, max_it = 100, n_h = 15,
  tolerance_intrinsic = 10^(-6), X = NULL, X_new = NULL,
  create_pdf_vario = FALSE, pdf_parameters = NULL, metric_manifold,
  model_ts, vario_model, distance)
}
\arguments{
\item{data_coords}{\code{N*2} or \code{N*3} matrix of [lat,long], [x,y] or [x,y,z] coordinates. [lat,long] are supposed to
be provided in signed decimal degrees}

\item{data_val}{array [\code{p,p,N}] of \code{N} symmetric positive definite matrices of dimension \code{p*p}}

\item{K}{number of neighborhood (i.e., centers) to sample at each iteration}

\item{grid}{prediction grid}

\item{nk_min}{minimum number of observations within a neighborhood}

\item{B}{number of \texit{divide} iterations to perform}

\item{suppressMes}{\{\code{TRUE}, \code{FALSE}\} controls the level of interaction and warnings given}

\item{ker.width.intrinsic}{parameter controlling the width of the Gaussian kernel for the computation of the local mean (if 0, no
kernel is used)}

\item{graph.distance.complete}{\code{N*N} distance matrix (the [i,j] element is the length of the shortest path between points i and j)}

\item{data.grid.distance}{\code{N*dim(grid)[1]} distance matrix between locations where the datum has been observed and locations where}

\item{N_samples}{number of samples}

\item{p}{dimension of the manifold matrices}

\item{aggregation_mean}{"Weighted" if the prediction obtained using the intrinsic mean, must be aggregated using different weights, "Equal" to use equal weights}

\item{metric_ts}{metric used on the tangent space. It must be chosen among "Frobenius", "FrobeniusScaled", "Correlation"}

\item{tol}{tolerance for the main loop of model_kriging}

\item{max_it}{maximum number of iterations for the main loop of model_kriging}

\item{n_h}{number of bins in the empirical variogram}

\item{tolerance_intrinsic}{tolerance for the computation of the intrinsic mean. Not needed if Sigma is provided}

\item{X}{matrix (N rows and unrestricted number of columns) of additional covariates for the tangent space model, possibly NULL}

\item{X_new}{matrix (with the same number of rows of \code{new_coords}) of additional covariates for the new locations, possibly NULL}

\item{create_pdf_vario}{boolean. If \code{TRUE} the empirical and fitted variograms are plotted in a pdf file}

\item{pdf_parameters}{list with the fields \code{test_nr} and \code{sample_draw}. Additional parameters to name the pdf}

\item{metric_manifold}{metric used on the manifold. It must be chosen among "Frobenius", "LogEuclidean", "SquareRoot"}

\item{model_ts}{type of model fitted on the tangent space. It must be chosen among "Intercept", "Coord1", "Coord2", "Additive"}

\item{vario_model}{type of variogram fitted. It must be chosen among "Gaussian", "Spherical", "Exponential"}

\item{distance}{type of distance between coordinates. It must be either "Eucldist" or "Geodist"}
}
\value{
it returns a list with the following fields
\itemize{
      \item \code{resBootstrap}{...}
      \item \code{resAggregated}{...}
      \item \code{model_pred}{...}
}
}
\description{
It employs a \emph{divide} et \emph{impera} strategy only to provide an estimate of a "fictional" field of tangent 
points, used to encode the information regarding the drift. To this end in the \emph{divide} step, the domain is randomly 
decomposed and in each subdomain a tangent point (assigned to each location in that subregion) is estimated as the 
intrinsic mean of the data belonging toit. This is repeated \code{K} times with different partitions of the domain and the 
results are then aggregated in the \emph{impera} stage by means of the intrinsic mean. Eventually, 
exploiting this "fictional" field of tangent points and the concept of parallel transport, a kriging analysis over the whole 
domain is performed to predict the field values at new locations.
}
\details{
The manifold values are mapped on the tangent space and then a GLS model is fitted to them. A first estimate of the beta coefficients
is obtained assuming spatially uncorrelated errors. Then, in the main the loop, new estimates of the beta are obtained as a result of a
weighted least square problem where the weight matrix is the inverse of \code{gamma_matrix}. The residuals \cr
\code{(residuals = data_ts - fitted)}
are updated accordingly. The parameters of the variogram fitted to the residuals (and used in the evaluation of the \code{gamma_matrix}) are
computed using Gauss-Newton with backtrack method to solve the associated non-linear least square problem. The stopping criteria is based on the
absolute value of the variogram residuals' norm if \code{ker.width.vario=0}, while it is based on its increment otherwise.
Once the model is computed, simple kriging on the tangent space is performed in correspondence of the new locations and eventually
the estimates are mapped to the manifold.
}
