% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_RDD.R
\name{mixed_RDD}
\alias{mixed_RDD}
\title{Perform mixed_RDD}
\usage{
mixed_RDD(data_coords, data_val, K, grid, nk_min = 1, B = 100,
  suppressMes = F, ker.width.intrinsic = 0, graph.distance.complete,
  data.grid.distance, N_samples, aggregation_mean, metric_ts,
  tol = 1e-12, max_it = 100, n_h = 15,
  tolerance_intrinsic = 10^(-6), max_sill = NULL, max_a = NULL,
  X = NULL, X_new = NULL, create_pdf_vario = FALSE,
  pdf_parameters = NULL, metric_manifold, model_ts, vario_model,
  distance)
}
\arguments{
\item{data_coords}{\code{N*2} or \code{N*3} matrix of [lat,long], [x,y] or [x,y,z] coordinates. [lat,long] are supposed to
be provided in signed decimal degrees}

\item{data_val}{array [\code{p,p,N}] of \code{N} symmetric positive definite matrices of dimension \code{p*p}}

\item{K}{number of cells the domain is subdivided in}

\item{grid}{prediction grid, i.e. \code{M*2} or \code{M*3} matrix of coordinates where to predict}

\item{nk_min}{minimum number of observations within a cell}

\item{B}{number of \emph{divide} iterations to perform}

\item{suppressMes}{\{\code{TRUE}, \code{FALSE}\} controls the level of interaction and warnings given}

\item{ker.width.intrinsic}{parameter controlling the width of the Gaussian kernel for the computation of the local mean (if 0, 
a "step kernel" is used, giving weight 1 to all the data within the cell and 0 to those outside of it)}

\item{graph.distance.complete}{\code{N*N} distance matrix (the [i,j] element is the length of the shortest path between points i and j)}

\item{data.grid.distance}{\code{N*M} distance matrix between locations where the datum has been observed and locations where
the datum has to be predicted}

\item{N_samples}{number of data N}

\item{aggregation_mean}{"Weighted" to aggregate the mean predictions using kernel-based weights, "Equal" to use equal weights}

\item{metric_ts}{metric used on the tangent space. It must be chosen among "Frobenius", "FrobeniusScaled", "Correlation"}

\item{tol}{tolerance for the main loop of \code{model_kriging}}

\item{max_it}{maximum number of iterations for the main loop of \code{model_kriging}}

\item{n_h}{number of bins in the empirical variogram}

\item{tolerance_intrinsic}{tolerance for the computation of the intrinsic mean}

\item{max_sill}{max value allowed for \code{sill} in the fitted variogram. If NULL it is defined as \code{1.15*max(emp_vario_values)}}

\item{max_a}{maximum value for \code{a} in the fitted variogram. If NULL it is defined as \code{1.15*h_max}}

\item{X}{matrix (N rows and unrestricted number of columns) of additional covariates for the tangent space model, possibly NULL}

\item{X_new}{matrix (with the same number of rows of \code{new_coords}) of additional covariates for the new locations, possibly NULL}

\item{create_pdf_vario}{boolean. If \code{TRUE} the empirical and fitted variograms are plotted in a pdf file}

\item{pdf_parameters}{list with the fields \code{test_nr} and \code{sample_draw}. Additional parameters to name the pdf}

\item{metric_manifold}{metric used on the manifold. It must be chosen among "Frobenius", "LogEuclidean", "SquareRoot"}

\item{model_ts}{type of model fitted on the tangent space. It must be chosen among "Intercept", "Coord1", "Coord2", "Additive"}

\item{vario_model}{type of variogram fitted. It must be chosen among "Gaussian", "Spherical", "Exponential"}

\item{distance}{type of distance between coordinates. It must be either "Eucldist" or "Geodist"}
}
\value{
it returns a list with the following fields
\itemize{
      \item \code{resBootstrap} {list of length \code{B}. Each field contains a tangent point estimate (at iteration \code{b}) for each new location, obtained 
                            as the intrinsic mean of the data within the tile it belongs to}
      \item \code{resAggregated} {field of tangent points computed, for each location (both those where data are measured and where they must be predicted), aggregating the corresponding \code{resBootstrap}}
      \item \code{model_pred} {list with the details of the global model fitted on the common Hibert space and the resulting kriging predictions. Namely it contains the following fields:
              \item{\code{beta}}{ vector of the beta matrices of the fitted model}
              \item{\code{gamma_matrix}}{ \code{N*N} covariogram matrix}
              \item{\code{residuals}}{ vector of the \code{N} residual matrices}
              \item{\code{emp_vario_values}}{ vector of empircal variogram values in correspondence of \code{h_vec}}
              \item{\code{h_vec}}{ vector of positions at which the empirical variogram is computed}
              \item{\code{fitted_par_vario}}{ estimates of \emph{nugget}, \emph{sill-nugget} and \emph{practical range}}
              \item{\code{iterations}}{ number of iterations of the main loop}
              \item{\code{prediction}}{ vector of matrices predicted at the new locations}}
}
}
\description{
Perform kriging prediction using MixedRDD procedure
}
\details{
It employs a \emph{divide} et \emph{impera} strategy to provide an estimate of a "fictional" field of tangent 
points, used to encode the information regarding the drift of the field. To this end in the \emph{divide} step, the domain is randomly 
decomposed and in each subdomain a tangent point (assigned to each location in that subregion) is estimated as the 
intrinsic mean of the data belonging to it. This is repeated \code{B} times with different partitions of the domain and the 
results are then aggregated in the \emph{impera} stage by means of the intrinsic mean. Eventually, 
exploiting this "fictional" field of tangent points and the concept of parallel transport, a kriging analysis over the whole 
domain is performed to predict the field values at new locations.
}
