% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_RDD.R
\name{mixed_RDD}
\alias{mixed_RDD}
\title{Perform mixed_RDD}
\usage{
mixed_RDD(data_coords, data_val, K, grid, nk_min = 1, B = 100,
  suppressMes = F, ker.width.intrinsic = 0, graph.distance.complete,
  data.grid.distance, N_samples, p, num.signif.entries, aggregation_mean,
  metric_ts, tol = 1e-12, max_it = 100, n_h = 15,
  tolerance_intrinsic = 10^(-6), X = NULL, X_new = NULL,
  X_tot = NULL, plot = FALSE, metric_manifold, model_ts, vario_model,
  distance)
}
\description{
Perform mixed_RDD
}
