% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_GLS.R
\name{model_GLS}
\alias{model_GLS}
\title{Create a GLS model}
\usage{
model_GLS(data_manifold, coords, X = NULL, Sigma = NULL,
  metric_manifold = "Frobenius", metric_ts = "Frobenius",
  model_ts = "Additive", vario_model = "Gaussian", n_h = 15,
  distance = "Geodist", max_it = 100, tolerance = 1e-06,
  weight_intrinsic = NULL, tolerance_intrinsic = 1e-06,
  max_sill = NULL, max_a = NULL, param_weighted_vario = NULL,
  plot = FALSE, suppressMes = FALSE, weight_extrinsic = NULL,
  tolerance_map_cor = 1e-06)
}
\arguments{
\item{data_manifold}{list or array [\code{p,p,N}] of \code{N} symmetric positive definite matrices of dimension \code{p*p}}

\item{coords}{\code{N*2} or \code{N*3} matrix of [lat,long], [x,y] or [x,y,z] coordinates. [lat,long] are supposed to
be provided in signed decimal degrees}

\item{X}{matrix (N rows and unrestricted number of columns) of additional covariates for the tangent space model, possibly NULL}

\item{Sigma}{\code{p*p} matrix representing the tangent point. If NULL the tangent point is computed as the intrinsic mean
of \code{data_manifold}}

\item{metric_manifold}{metric used on the manifold. It must be chosen among "Frobenius", "LogEuclidean", "SquareRoot", "Correlation"}

\item{metric_ts}{metric used on the tangent space. It must be chosen among "Frobenius", "FrobeniusScaled", "Correlation"}

\item{model_ts}{type of model fitted on the tangent space. It must be chosen among "Intercept", "Coord1", "Coord2", "Additive"}

\item{vario_model}{type of variogram fitted. It must be chosen among "Gaussian", "Spherical", "Exponential"}

\item{n_h}{number of bins in the emprical variogram}

\item{distance}{type of distance between coordinates. It must be either "Eucldist" or "Geodist"}

\item{max_it}{max number of iterations for the main loop}

\item{tolerance}{tolerance for the main loop}

\item{weight_intrinsic}{vector of length \code{N} to weight the locations in the computation of the intrinsic mean. If NULL
a vector of ones is used. Not needed if \code{Sigma} is provided}

\item{tolerance_intrinsic}{tolerance for the computation of the intrinsic mean. Not needed if Sigma is provided}

\item{max_sill}{max value allowed for \emph{sill} in the fitted variogram. If NULL it is defined as \code{1.15*max(emp_vario_values)}}

\item{max_a}{maximum value for \emph{a} in the fitted variogram. If NULL it is defined as \code{1.15*h_max}}

\item{param_weighted_vario}{List of 7 elements to be provided to consider Kernel weights for the variogram (significant only within an RDD procedure). 
Indeed in this case the N_tot data regarding the whole domain must be provided to the algorithm, not only the N in the cell under consideration. Therefore 
the list must contain the following fields:
\code{weight_vario} (vector of length \code{N_tot} to weight the locations in the computation of the empirical variogram),
\code{distance_matrix_tot} (\code{N_tot*N_tot} matrix of distances between the locations), \code{data_manifold_tot}
(list or array \cr [\code{p,p,N_tot}] of \code{N_tot} symmetric positive definite matrices of dimension \code{p*p}),
\code{coords_tot} (\code{N_tot*2} or \code{N_tot*3} matrix of [lat,long], [x,y] or [x,y,z] coordinates),
\code{X_tot} (matrix with N_tot rows and unrestricted number of columns of additional covariates for the tangent space model, possibly NULL),
\code{h_max} (maximum value of distance for which the variogram is computed), 
\code{indexes_model} (indexes of the N_tot data corresponding to the N data in the cell).}

\item{plot}{boolean. If \code{TRUE} the empirical and fitted variograms are plotted}

\item{suppressMes}{boolean. If \code{TRUE} warning messagges are not printed}

\item{weight_extrinsic}{vector of length \code{N} to weight the locations in the computation of the extrinsic mean. If NULL
weight_intrinsic are used. Needed only if \code{Sigma} is not provided and \code{metric_manifold== "Correlation"}}

\item{tolerance_map_cor}{tolerance to use in the maps. \cr Required only if
\code{metric_manifold== "Correlation"}}

\item{data_dist_mat}{Matrix of dimension \code{N*N} of distances between data points. If not provided it is computed using \code{distance}}
}
\value{
A list with the following fields:
\item{\code{beta}}{ vector of the beta matrices of the fitted model}
\item{\code{gamma_matrix}}{ \code{N*N} covariogram matrix}
\item{\code{residuals}}{ vector of the \code{N} residual matrices}
\item{\code{emp_vario_values}}{ vector of empircal variogram values in correspondence of \code{h_vec}}
\item{\code{h_vec}}{ vector of positions at which the empirical variogram is computed}
\item{\code{fitted_par_vario}}{ estimates of \emph{nugget}, \emph{sill-nugget} and \emph{practical range}}
\item{\code{iterations}}{ number of iterations of the main loop}
\item{\code{Sigma}}{ tangent point}
}
\description{
Given the coordinates and corresponding manifold values, this function creates a GLS model on the tangent space.
}
\details{
The manifold values are mapped on the tangent space and then a GLS model is fitted to them. A first estimate of the \code{beta} coefficients
is obtained assuming spatially uncorrelated errors. Then, in the main the loop, new estimates of the \code{beta} are obtained as a result of a
weighted least square problem where the weight matrix is the inverse of \code{gamma_matrix}. The residuals
\cr\code{(residuals = data_ts - fitted)}
are updated accordingly. The parameters of the variogram fitted to the residuals (and used in the evaluation of the \code{gamma_matrix}) are
computed using Gauss-Newton with backtrack method to solve the associated non-linear least square problem. The stopping criteria is based on the
absolute value of the variogram residuals' norm if \code{ker.width.vario=0}, while it is based on its increment otherwise.
}
\examples{
data_manifold_model <- Manifoldgstat::rCov
coords_model <- Manifoldgstat::rGrid
Sigma <- matrix(c(2,1,1,1), 2,2)
model = model_GLS(data_manifold = data_manifold_model, coords = coords_model, Sigma = Sigma,
                 metric_manifold = "Frobenius", metric_ts = "Frobenius", model_ts = "Coord1",
                 vario_model = "Spherical", n_h = 15, distance = "Eucldist", max_it = 100,
                 tolerance = 1e-7, plot = TRUE)
}
\references{
D. Pigoli, A. Menafoglio & P. Secchi (2016):
Kriging prediction for manifold-valued random fields.
Journal of Multivariate Analysis, 145, 117-131.
}
