% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_GLS.R
\name{model_GLS}
\alias{model_GLS}
\title{Create a GLS model}
\usage{
model_GLS(data_manifold, coords, X = NULL, Sigma = NULL,
  metric_manifold = "Frobenius", metric_ts = "Frobenius",
  model_ts = " additive", vario_model = "Gaussian", n_h = 15,
  distance = "Geodist", max_it = 100, tolerance = 1e-06,
  weight_vario = NULL, weight_intrinsic = NULL,
  tolerance_intrinsic = 1e-06, plot = FALSE)
}
\arguments{
\item{data_manifold}{list or array [\code{n,n,N}] of \code{N} symmetric positive definite matrices of dimension \code{nxn}}

\item{coords}{\code{N*2} or \code{N*3} matrix of [lat,long], [x,y] or [x,y,z] coordinates. [lat,long] are supposed to
be provided in signed decimal degrees}

\item{X}{matrix (N rows and unrestricted number of columns) of additional covariates for the tangent space model, possibly NULL}

\item{Sigma}{\code{n*n} matrix representing the tangent point. If NULL the tangent point is computed as the intrinsic mean
of \code{data_manifold}}

\item{metric_manifold}{metric used on the manifold. It must be chosen among "Frobenius", "LogEuclidean", "SquareRoot"}

\item{metric_ts}{metric used on the tangent space. It must be either "Frobenius" or "FrobeniusScaled"}

\item{model_ts}{type of model fitted on the tangent space. It must be chosen among "Intercept", "Coord1", "Coord2", "Additive"}

\item{vario_model}{type of variogram fitted. It must be chosen among "Gaussian", "Spherical", "Exponential"}

\item{n_h}{number of bins in the emprical variogram}

\item{distance}{type of distance between coordinates. It must be either "Eucldist" or "Geodist"}

\item{max_it}{max number of iterations for the main loop}

\item{tolerance}{tolerance for the main loop}

\item{weight_vario}{vector of length \code{N} to weight the locations in the computation of the empirical variogram. If NULL
a vector of ones is used}

\item{weight_intrinsic}{vector of length \code{N} to weight the locations in the computation of the intrinsic mean. If NULL
a vector of ones is used. Not needed if Sigma is provided}

\item{tolerance_intrinsic}{tolerance for the computation of the intrinsic mean. Not needed if Sigma is provided}

\item{plot}{boolean. If \code{TRUE} the empirical and fitted variograms are plotted}
}
\value{
A \code{list} with the following fields:
\item{\code{beta}}{vector of the beta matrices of the fitted model}
\item{\code{gamma_matrix}}{\code{N*N} covariogram matrix}
\item{\code{residuals}}{vector of the \code{N} residual matrices}
\item{\code{emp_vario_values}}{vector of empircal variogram values in correspondence of \code{h_vec}}
\item{\code{h_vec}}{vector of positions at which the empirical variogram is computed}
\item{\code{fitted_par_vario}}{estimates of \code{nugget}, \code{sill-nugget} and \code{practical range}}
\item{\code{iterations}}{number of iterations of the main loop}
\item{\code{Sigma}}{tangent point}
}
\description{
Given the coordinates and corresponding manifold values, this function creates a GLS model on the tangent space.
}
\details{
The manifold values are mapped on the tangent space and then a GLS model is fitted to them. A first estimate of the beta coefficients 
is obtained assuming spatially uncorrelated errors. Then, in the main the loop, new estimates of the beta are obtained as a result of a 
weighted least square problem where the weight matrix is the inverse of \code{gamma_matrix}. The residuals \code{(residuals = data_ts - fitted)} 
are updated accordingly. The parameters of the variogram fitted to the residuals (and used in the evaluation of the \code{gamma_matrix}) are 
computed using Gauss-Newton with backtrack method to solve the associated non-linear least square problem.
}
\examples{
data_manifold_model <- Manifoldgstat::rCov
coords_model <- Manifoldgstat::rGrid
Sigma <- matrix(c(2,1,1,1), 2,2)
model = model_GLS(data_manifold = data_manifold_model, coords = coords_model, Sigma = Sigma, metric_manifold = "Frobenius",
                   metric_ts = "Frobenius", model_ts = "Coord1", vario_model = "Spherical", n_h = 15, 
                   distance = "Eucldist", max_it = 100, tolerance = 1e-7, plot = TRUE)
}
\references{
D. Pigoli, A. Menafoglio & P. Secchi (2016):
Kriging prediction for manifold-valued random fields.
Journal of Multivariate Analysis, 145, 117-131.
}
