% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_kriging.R
\name{model_kriging}
\alias{model_kriging}
\title{Model and Kriging together}
\usage{
model_kriging(data_manifold, coords, X = NULL, Sigma,
  metric_manifold = "Frobenius", metric_ts = "Frobenius",
  model_ts = "additive", vario_model = "Gaussian", n_h = 15,
  distance = "Geodist", max_it = 100, tolerance = 10^(-4),
  weight = NULL, new_coords, X_new = NULL)
}
\description{
Model and Kriging together
}
